+++ CHANGELOG.md
@@ -91,3 +91,7 @@
 * Better handling of errors (i.e. DON'T IGNORE ERRORS) when MSMQ receive fails
 * Optimization of SQL Server transport
 * Fixed hard-to-find bug in how `SqlServerMessageQueue` would associate commands with the ongoing `SqlTransaction`
+
+## 0.35.3
+
+* Added transport performance showdown